Depuracion

+++ Versiones/proyecto/trazabilidad-ejb/src/main/java/ejb/CuentaRefEJB.java
@@ -10,8 +10,12 @@ import javax.persistence.TypedQuery;
 import exceptions.CuentaExistenteException;
 import exceptions.CuentaNoEncontradoException;
 import exceptions.ProyectoException;
+import exceptions.UserExistenteException;
+import exceptions.UserNoAdminException;
+import exceptions.UserNoEncontradoException;
 import jpa.Cuenta;
 import jpa.CuentaRef;
+import jpa.UserApk;
 
 @Stateless
 public class CuentaRefEJB implements GestionCuentaRef {
@@ -20,13 +24,22 @@ public class CuentaRefEJB implements GestionCuentaRef {
 	private EntityManager em;
     
 	@Override
-	public void insertarCuentaRef(CuentaRef cuenta) throws CuentaExistenteException {
-		CuentaRef cuentaExistente = em.find(CuentaRef.class, cuenta.getIBAN());
-		if (cuentaExistente != null) {
-			throw new CuentaExistenteException();
+	public void insertarCuentaRef(UserApk user, CuentaRef cuenta) throws ProyectoException {
+		UserApk userExistente= em.find(UserApk.class, user.getUser());
+		if (userExistente == null) {
+			throw new UserNoEncontradoException();
+		} else {
+			if(user.isAdministrativo()) {
+				CuentaRef cuentaExistente = em.find(CuentaRef.class, cuenta.getIBAN());
+				if (cuentaExistente != null) {
+					throw new CuentaExistenteException();
+				}
+				
+				 em.persist(cuenta);
+			} else {
+				throw new UserNoAdminException();
+			}
 		}
-		
-		 em.persist(cuenta);
 	}
 
 	@Override
@@ -36,48 +49,71 @@ public class CuentaRefEJB implements GestionCuentaRef {
 	}
 
 	@Override
-	public void actualizarCuentaRef(CuentaRef cuenta) throws ProyectoException {
-		CuentaRef cuentaEntity = em.find(CuentaRef.class, cuenta.getIBAN());
-		if (cuentaEntity == null) {
-			throw new CuentaNoEncontradoException();
+	public void actualizarCuentaRef(UserApk user, CuentaRef cuenta) throws ProyectoException {
+		UserApk userExistente= em.find(UserApk.class, user.getUser());
+		if (userExistente == null) {
+			throw new UserNoEncontradoException();
+		} else {
+			if(user.isAdministrativo()) {
+				CuentaRef cuentaEntity = em.find(CuentaRef.class, cuenta.getIBAN());
+				if (cuentaEntity == null) {
+					throw new CuentaNoEncontradoException();
+				}
+				
+				cuentaEntity.setNombreBanco(cuenta.getNombreBanco());
+				cuentaEntity.setSucursal(cuenta.getSucursal());
+				cuentaEntity.setPais(cuenta.getPais());
+				cuentaEntity.setSaldo(cuenta.getSaldo());
+				cuentaEntity.setFechaApertura(cuenta.getFechaApertura());
+				cuentaEntity.setEstado(cuenta.getEstado());
+				
+				em.merge(cuentaEntity);
+			} else {
+				throw new UserNoAdminException();
+			}
 		}
-		
-		cuentaEntity.setNombreBanco(cuenta.getNombreBanco());
-		cuentaEntity.setSucursal(cuenta.getSucursal());
-		cuentaEntity.setPais(cuenta.getPais());
-		cuentaEntity.setSaldo(cuenta.getSaldo());
-		cuentaEntity.setFechaApertura(cuenta.getFechaApertura());
-		cuentaEntity.setEstado(cuenta.getEstado());
-		
-		em.merge(cuentaEntity);
 	}
 
 	@Override
-	public void eliminarCuentaRef(CuentaRef cuenta) throws ProyectoException {
-		CuentaRef cuentaRefEntity = em.find(CuentaRef.class, cuenta.getIBAN());
-		Cuenta cuentaEntity = em.find(Cuenta.class, cuenta.getIBAN());
-		
-		if (cuentaRefEntity == null && (cuentaEntity == null)) {
-			throw new CuentaNoEncontradoException();
+	public void eliminarCuentaRef(UserApk user, CuentaRef cuenta) throws ProyectoException {
+		UserApk userExistente= em.find(UserApk.class, user.getUser());
+		if (userExistente == null) {
+			throw new UserNoEncontradoException();
+		} else {
+			if(user.isAdministrativo()) {
+				CuentaRef cuentaRefEntity = em.find(CuentaRef.class, cuenta.getIBAN());
+				
+				if (cuentaRefEntity == null) {
+					throw new CuentaNoEncontradoException();
+				}
+				
+				em.remove(cuentaRefEntity);
+			} else {
+				throw new UserNoAdminException();
+			}
 		}
-		
-		em.remove(cuentaRefEntity);
-		em.remove(cuentaEntity);
-		
 	}
 
 	@Override
-	public void eliminarTodasCuentasRef() throws ProyectoException {
-		List<CuentaRef> cuentas = obtenerCuentasRef();
-		
-		for (CuentaRef e : cuentas) {
-			Cuenta cuentaEntity = em.find(Cuenta.class, e.getIBAN());
-			em.remove(cuentaEntity);
-		}
-		
-		for (CuentaRef e : cuentas) {
-			em.remove(e);
-		}
-		
+	public void eliminarTodasCuentasRef(UserApk user) throws ProyectoException {
+		UserApk userExistente= em.find(UserApk.class, user.getUser());
+		if (userExistente == null) {
+			throw new UserNoEncontradoException();
+		} else {
+			if(user.isAdministrativo()) {
+				List<CuentaRef> cuentas = obtenerCuentasRef();
+				
+				for (CuentaRef e : cuentas) {
+					Cuenta cuentaEntity = em.find(Cuenta.class, e.getIBAN());
+					em.remove(cuentaEntity);
+				}
+				
+				for (CuentaRef e : cuentas) {
+					em.remove(e);
+				}	
+			} else {
+				throw new UserNoAdminException();
+			}
+		}	
 	}
 }
