+++ Versiones/proyecto/trazabilidad-ejb/src/main/java/ejb/TransEJB.java
@@ -8,7 +8,10 @@ import javax.persistence.TypedQuery;
 import exceptions.ProyectoException;
 import exceptions.TransExistenteException;
 import exceptions.TransNoEncontradaException;
+import exceptions.UserNoAdminException;
+import exceptions.UserNoEncontradoException;
 import jpa.Trans;
+import jpa.UserApk;
 
 @Stateless
 public class TransEJB implements GestionTrans{
@@ -33,44 +36,69 @@ public class TransEJB implements GestionTrans{
 		}
 
 		@Override
-		public void actualizarTrans(Trans trans) throws ProyectoException {
-			Trans transEntity = em.find(Trans.class, trans.getID());
-			if (transEntity == null) {
-				throw new TransNoEncontradaException();
-			}
+		public void actualizarTrans(UserApk user, Trans trans) throws ProyectoException {
+			UserApk userEntity = em.find(UserApk.class, user.getUser());
+			if (userEntity == null) {
+				throw new UserNoEncontradoException();
+			} else {
+				if(user.isAdministrativo()) {
+					Trans transEntity = em.find(Trans.class, trans.getID());
+					if (transEntity == null) {
+						throw new TransNoEncontradaException();
+					}
 
-			transEntity.setCantidad(trans.getCantidad());
-			transEntity.setTipo(trans.getTipo());
-			transEntity.setComision(trans.getComision());
-			transEntity.setInternational(trans.isInternational());
-			transEntity.setFechaInstruccion(trans.getFechaInstruccion());
-			transEntity.setFechaEjecucion(trans.getFechaEjecucion());
-			transEntity.setMonedaDestino(trans.getMonedaOrigen());
-			transEntity.setMonedaDestino(trans.getMonedaDestino());
-			transEntity.setCuenta(trans.getCuenta());
-			transEntity.setTransaccion(trans.getTransaccion());
-			
-			em.merge(transEntity);
+					transEntity.setCantidad(trans.getCantidad());
+					transEntity.setTipo(trans.getTipo());
+					transEntity.setComision(trans.getComision());
+					transEntity.setInternational(trans.isInternational());
+					transEntity.setFechaInstruccion(trans.getFechaInstruccion());
+					transEntity.setFechaEjecucion(trans.getFechaEjecucion());
+					transEntity.setMonedaDestino(trans.getMonedaOrigen());
+					transEntity.setMonedaDestino(trans.getMonedaDestino());
+					transEntity.setCuenta(trans.getCuenta());
+					transEntity.setTransaccion(trans.getTransaccion());
+					
+					em.merge(transEntity);
+				} else {
+					throw new UserNoAdminException();
+				}
+			}
 		}
 
 		@Override
-		public void eliminarTrans(Trans trans) throws ProyectoException {
-			Trans transEntity = em.find(Trans.class, trans.getID());
-			if (transEntity == null) {
-				throw new TransNoEncontradaException();
+		public void eliminarTrans(UserApk user, Trans trans) throws ProyectoException {
+			UserApk userEntity = em.find(UserApk.class, user.getUser());
+			if (userEntity == null) {
+				throw new UserNoEncontradoException();
+			} else {
+				if(user.isAdministrativo()) {
+					Trans transEntity = em.find(Trans.class, trans.getID());
+					if (transEntity == null) {
+						throw new TransNoEncontradaException();
+					}
+					
+					em.remove(transEntity);
+				} else {
+					throw new UserNoAdminException();
+				}
 			}
-			
-			em.remove(transEntity);
-			
 		}
 
 		@Override
-		public void eliminarTodasTrans() throws ProyectoException {
-			List<Trans> trans = obtenerTrans();
-			
-			for (Trans t : trans) {
-				em.remove(t);
+		public void eliminarTodasTrans(UserApk user) throws ProyectoException {
+			UserApk userEntity = em.find(UserApk.class, user.getUser());
+			if (userEntity == null) {
+				throw new UserNoEncontradoException();
+			} else {
+				if(user.isAdministrativo()) {
+					List<Trans> trans = obtenerTrans();
+					
+					for (Trans t : trans) {
+						em.remove(t);
+					}
+				} else {
+					throw new UserNoAdminException();
+				}
 			}
-			
 		}
 }
